+++ wordpressinstall.sh
@@ -77,9 +77,11 @@ fi
 
 #php 7.2
 echo '{"pkgs":["nano","rsync","nginx","mariadb102-server","php72","php72-mysqli","php72-session","php72-xml","php72-hash","php72-ftp","php72-curl","php72-tokenizer","php72-zlib","php72-zip","php72-filter","php72-gd","php72-openssl"]}' > /tmp/pkg.json
-
-iocage create --name "${JAIL_NAME}" -p /tmp/pkg.json -r $RELEASE ip4_addr="${INTERFACE}|${JAIL_IP}/24" defaultrouter="${DEFAULT_GW_IP}" boot="on" host_hostname="${JAIL_NAME}" vnet="${VNET}"
-
+if ! iocage create --name "${JAIL_NAME}" -p /tmp/pkg.json -r "${RELEASE}" ip4_addr="${INTERFACE}|${JAIL_IP}/24" defaultrouter="${DEFAULT_GW_IP}" boot="on" host_hostname="${JAIL_NAME}" vnet="${VNET}"
+then
+	echo "Failed to create jail"
+	exit 1
+fi
 rm /tmp/pkg.json
 
 mkdir -p ${POOL_PATH}/${APPS_PATH}/${WP_DATA}
